Owe the name lock to a CLEARED name as well as a supplied one
Two channels write an organization's name and only one was ever asked
about. `Clear: ["legal_name"]` sets the column to NULL, lands in the
patch's after-image and triggers the duplicate re-check exactly as a
rename does — but it skipped the name lock entirely, so it took the row
lock through the guarded write and then reached for the name lock inside
the re-check. That is the ordering renamerecheck.go forbids, and it is
the serialization the re-check depends on.

The question is now asked in one place, against one named set of name
columns that the re-check reads too, and the cleared wire field is
resolved onto its column through the map that decides what is clearable
at all — so a name that becomes clearable later is covered on the day it
does.

The edit path is split out of organization.go behind that: the update
grew past both the file's length cap and its own complexity budget, so
the transaction body, the record it leaves, the conditional name lock
and the geocode enqueue each have a name. The provenance stamp stays
inline with the guarded write it rides — moved into a helper it leaves
its concurrency guard and its write-authority probe behind, and the two
gates that check for them are right to say so.
