+++ backend/internal/modules/people/organization.go
@@ -6,7 +6,6 @@ package people
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/jackc/pgx/v5"
 
@@ -202,258 +201,4 @@ func createOrganizationInTx(ctx context.Context, tx pgx.Tx, in CreateOrganizatio
 		return crmcontracts.Organization{}, fmt.Errorf("read created organization: %w", err)
 	}
 	return out, nil
-}
-
-type UpdateOrganizationInput struct {
-	// Clear names the wire fields to set to NULL. A JSON null cannot say so —
-	// it decodes to a nil pointer and reads as "not supplied" — so the
-	// reversal path names them here instead.
-	Clear []string
-	// Trail names what the audit trail calls this write; zero is an update.
-	Trail       storekit.AuditTrail
-	DisplayName *string
-	LegalName   *string
-	// Description, when non-nil, sets or (when empty) clears the one-line
-	// summary the company page shows under the title. nil leaves it untouched.
-	Description *string
-	Industry    *string
-	SizeBand    *string
-	OwnerID     *ids.UserID
-	ParentOrgID *ids.OrganizationID
-	Address     *crmcontracts.Address
-	IfVersion   *int64
-	// LinkedInURL, when non-nil, sets or (when empty) clears the canonical
-	// LinkedIn company URL (PO-DDL-N-2). nil leaves it untouched.
-	LinkedInURL *string
-	// Domains, when non-nil, is the desired live domain set (replace-set:
-	// add missing, archive removed, flip is_primary). nil leaves domains
-	// untouched; an empty slice clears them.
-	Domains *[]OrgDomainInput
-	// Lifecycle, when non-nil, moves where the account stands with us
-	// (ADR-0079/A124). nil leaves it untouched.
-	Lifecycle *string
-	// RelationshipTypes, when non-nil, is the desired live type set — the same
-	// replace-set shape as Domains. nil leaves them untouched; an empty slice
-	// clears them, except that 'partner' cannot be dropped while the partner
-	// extension row lives.
-	RelationshipTypes *[]string
-	// CustomFields carries the request body's extra top-level keys
-	// (additionalProperties); only active cf_* catalog columns land,
-	// drop-on-mismatch (customfields.go).
-	CustomFields map[string]any
-}
-
-func (s *Store) UpdateOrganization(ctx context.Context, id ids.OrganizationID, in UpdateOrganizationInput) (crmcontracts.Organization, error) {
-	if err := auth.Require(ctx, "organization", principal.ActionUpdate); err != nil {
-		return crmcontracts.Organization{}, err
-	}
-	active, err := s.activeColumns(ctx, "organization")
-	if err != nil {
-		return crmcontracts.Organization{}, err
-	}
-	var out crmcontracts.Organization
-	err = s.tx(ctx, func(tx pgx.Tx) error {
-		if err := auth.EnsureWritable(ctx, tx, "organization", id.UUID); err != nil {
-			return err
-		}
-		// Only a rename needs the name lock, and only a rename should pay for
-		// it: the key is workspace-wide, so taking it for an owner change would
-		// serialize every organization write behind an edit that cannot create
-		// a duplicate. Taken ahead of the patch's row lock, per the ordering
-		// rule on lockOrgNameWrites.
-		//
-		// And ahead of the READ, because the read is what the rename is judged
-		// against. Taken after it, a concurrent rename landing in between left
-		// `current` holding a name nobody stored any more — so this write
-		// overwrote that rename while its before-image said the name had not
-		// moved, and the provenance stamp below was skipped for an edit that
-		// really did change the name.
-		if in.DisplayName != nil || in.LegalName != nil {
-			if err := lockOrgNameWrites(ctx, tx); err != nil {
-				return err
-			}
-		}
-		current, err := readOrganization(ctx, tx, id, storekit.LiveOnly, active)
-		if err != nil {
-			return fmt.Errorf("read organization before update: %w", err)
-		}
-		p, err := buildOrganizationPatch(ctx, tx, current, in)
-		if err != nil {
-			return err
-		}
-		storekit.SetCustomFieldPatch(p, active, in.CustomFields, current.AdditionalProperties)
-
-		by, err := stageOrgReplaceSets(ctx, tx, id, current, in, p)
-		if err != nil {
-			return err
-		}
-		if p.Empty() {
-			out = current
-			return nil
-		}
-		if err := p.ApplyGuarded(ctx, tx, "organization", id.UUID, in.IfVersion); err != nil {
-			return fmt.Errorf("apply organization patch: %w", err)
-		}
-		// INVALIDATION IS THE TRIGGER'S JOB, not this call's — the schema marks
-		// the coordinates stale on any address column that actually changed, so
-		// no writer has to remember and none can forget. What is left here is
-		// the enqueue, and it is keyed off what the patch actually MOVED rather
-		// than off the request carrying an address field: re-submitting a form
-		// with an unchanged address would otherwise spend a lookup, and every
-		// lookup is fifteen seconds of a rate the whole installation shares.
-		if movedAddress(p.After()) {
-			if err := s.enqueueGeocode(ctx, tx, id); err != nil {
-				return fmt.Errorf("re-locating a moved company: %w", err)
-			}
-		}
-
-		before, after := p.Before(), p.After()
-		// MOVED, not merely set. A human editing the display name is the top of
-		// the name-source lattice (ADR-0072/A118): stamp 'human' so no
-		// automated source ever overwrites it. Re-sending the same value is not
-		// a re-authoring, and reading After() made it one — an agent
-		// round-tripping a record it had just read froze a provisional
-		// domain-derived name for good, with nothing in the record saying a
-		// person had never chosen it. It rides the patch's own guarded write
-		// above, on the row that write just locked.
-		if _, changed := p.Moved()["display_name"]; changed {
-			if _, err := tx.Exec(ctx, `UPDATE organization SET name_source = 'human' WHERE id = $1`, id); err != nil {
-				return fmt.Errorf("stamp organization name provenance: %w", err)
-			}
-		}
-		// The description carries the same lattice on a different layer. A site
-		// read may replace a description no person authored, so an edited one has
-		// to say in field_provenance that a person wrote it — otherwise the next
-		// crawl reads "no row" as "no owner" and takes the sentence somebody
-		// typed.
-		//
-		// The test is whether the VALUE moved, not whether the field was sent.
-		// storekit.Patch records an assignment unconditionally, so `after` holds
-		// every key the request named — and an agent re-sending a human's
-		// description unchanged would otherwise write an agent: row on top of the
-		// human's and hand the column to the next crawl. The display-name stamp
-		// above now asks the same question through Patch.Moved. This one cannot:
-		// the description's two images hold different TYPES by construction —
-		// *string as read from the row, string as the request supplied it — and
-		// Moved counts a type change as a move, so it would call every send a
-		// rewrite. describedDifferently compares the values themselves.
-		if describedDifferently(before, after) {
-			// Not the `by` above: stageOrgReplaceSets answers "" unless the edit
-			// also carried a replace-set, and a description edit usually carries
-			// neither, which would stamp the field to nobody.
-			editor, err := storekit.CapturedBy(ctx)
-			if err != nil {
-				return err
-			}
-			if err := stampDescriptionAuthor(ctx, tx, id, editor); err != nil {
-				return err
-			}
-		}
-		if err := recheckRenamedOrganization(ctx, tx, id, after); err != nil {
-			return err
-		}
-		if err := reconcileOrgReplaceSets(ctx, tx, id, by, in, before, after); err != nil {
-			return err
-		}
-
-		auditID, err := storekit.AuditWithTrail(ctx, tx, in.Trail, "organization", id.UUID, before, after)
-		if err != nil {
-			return fmt.Errorf("audit organization update: %w", err)
-		}
-		if err := storekit.EmitEvent(ctx, tx, auditID, id.UUID, crmcontracts.PublicEventOrganizationUpdated{ChangedFields: after}); err != nil {
-			return fmt.Errorf("emit organization.updated: %w", err)
-		}
-		if out, err = readOrganization(ctx, tx, id, storekit.LiveOnly, active); err != nil {
-			return fmt.Errorf("read updated organization: %w", err)
-		}
-		return nil
-	})
-	return out, err
-}
-
-// stageOrgReplaceSets validates the domain and relationship-type replace-sets
-// up front, so a bad request fails before the version bump, and folds the
-// updated_at bump each of them rides into the patch: the reconciles happen
-// against the org row's own guarded write, so If-Match still guards them and
-// the audit row records the transition — the same shape as UpdatePerson/social.
-// It answers with the captured-by principal those reconciles stamp, empty when
-// the edit touches neither set.
-func stageOrgReplaceSets(ctx context.Context, tx pgx.Tx, id ids.OrganizationID,
-	current crmcontracts.Organization, in UpdateOrganizationInput, p *storekit.Patch,
-) (string, error) {
-	if in.RelationshipTypes == nil && in.Domains == nil {
-		return "", nil
-	}
-	by, err := storekit.CapturedBy(ctx)
-	if err != nil {
-		return "", err
-	}
-	if in.RelationshipTypes != nil {
-		deduped, err := dedupeRelationshipTypes(*in.RelationshipTypes)
-		if err != nil {
-			return "", err
-		}
-		*in.RelationshipTypes = deduped
-	}
-	if in.Domains != nil {
-		if err := parseOrgDomains(*in.Domains); err != nil {
-			return "", err
-		}
-		// Collapse hosts that normalize to the same domain so the probe,
-		// reconcile, and audit-after all see one row per host.
-		*in.Domains = dedupeDomains(*in.Domains)
-		if err := ensureOrgDomainsUnclaimedExcept(ctx, tx, id, *in.Domains); err != nil {
-			return "", err
-		}
-	}
-	// A replace-set changes no column on the row itself, so this bump is what
-	// makes the patch non-empty and carries the version guard. Once, after both
-	// branches: a request naming both sets is still one write at one instant, and
-	// which branch happened to run last should not pick the timestamp.
-	p.Set("updated_at", current.UpdatedAt, time.Now().UTC())
-	return by, nil
-}
-
-// reconcileOrgReplaceSets lands the two replace-sets on the row the patch just
-// wrote and records each transition in the audit images the caller commits.
-func reconcileOrgReplaceSets(ctx context.Context, tx pgx.Tx, id ids.OrganizationID, by string,
-	in UpdateOrganizationInput, before, after map[string]any,
-) error {
-	if in.Domains != nil {
-		domainsBefore, err := reconcileOrgDomains(ctx, tx, workspaceID(ctx), id, by, *in.Domains)
-		if err != nil {
-			return err
-		}
-		before["domains"] = domainsBefore
-		after["domains"] = domainSummaries(*in.Domains)
-	}
-	if in.RelationshipTypes != nil {
-		typesBefore, err := reconcileOrgRelationshipTypes(ctx, tx, workspaceID(ctx), id, "manual", by, *in.RelationshipTypes)
-		if err != nil {
-			return err
-		}
-		before["relationship_types"] = typesBefore
-		after["relationship_types"] = *in.RelationshipTypes
-	}
-	return nil
-}
-
-// recheckRenamedOrganization asks whether an edited organization now resembles
-// another one, given the patch's applied delta.
-//
-// Either name axis can reveal it alone: two records of one company converging
-// on the same registered name is exactly the shape that doubled a company in a
-// live workspace, and a legal-name-only edit changes no display name at all.
-// The edit stands regardless — this only files a pair for the review queue.
-func recheckRenamedOrganization(ctx context.Context, tx pgx.Tx, id ids.OrganizationID, after map[string]any) error {
-	_, renamedDisplay := after["display_name"]
-	_, renamedLegal := after["legal_name"]
-	if !renamedDisplay && !renamedLegal {
-		return nil
-	}
-	editor, err := storekit.CapturedBy(ctx)
-	if err != nil {
-		return err
-	}
-	return recheckOrgNameForDuplicates(ctx, tx, id, editor)
 }
