+++ backend/internal/modules/people/renamerecheck_test.go
@@ -7,6 +7,8 @@ import (
 	"strings"
 	"testing"
 	"time"
+
+	crmcontracts "github.com/margince/margince/backend/internal/contracts"
 )
 
 // The name shapes that put one company in a workspace twice, each pinned above
@@ -128,4 +130,64 @@ func TestAnEvidenceApplyOwesTheNameLockWheneverItCarriesANameField(t *testing.T)
 			}
 		})
 	}
+}
+
+// The same debt on the update path, and its second channel.
+//
+// An update writes a name two ways: it supplies one, or it CLEARS one. Both end
+// in the patch's after-image and both trigger the re-check, but only the
+// supplied value was ever asked about — so `Clear: ["legal_name"]` took the row
+// lock through the guarded write and then reached for the name lock inside the
+// re-check, which is the ordering lockOrgNameWrites exists to forbid.
+//
+// The clearable fields are read from the map that decides what is clearable at
+// all, rather than listed here: a name that becomes clearable later is covered
+// on the day it does, not on the day somebody remembers this test.
+func TestAnUpdateOwesTheNameLockThroughEitherChannelThatWritesAName(t *testing.T) {
+	t.Run("a supplied name", func(t *testing.T) {
+		display, legal := "Baqend", "Baqend GmbH"
+		for name, in := range map[string]UpdateOrganizationInput{
+			"display": {DisplayName: &display},
+			"legal":   {LegalName: &legal},
+		} {
+			if !renamesAnOrganization(in) {
+				t.Errorf("a supplied %s name owes the lock and was not asked for it", name)
+			}
+		}
+	})
+
+	t.Run("a cleared name", func(t *testing.T) {
+		names := 0
+		for field, clearable := range clearableOrganizationColumns(crmcontracts.Organization{}) {
+			owes := renamesAnOrganization(UpdateOrganizationInput{Clear: []string{field}})
+			if !orgNameColumns[clearable.Column] {
+				// The other end: the lock is workspace-wide, so a clear that
+				// writes no name must NOT pay for it.
+				if owes {
+					t.Errorf("clearing %q writes %s, which is no name, and it took the workspace-wide lock anyway",
+						field, clearable.Column)
+				}
+				continue
+			}
+			names++
+			if !owes {
+				t.Errorf("clearing %q writes %s and owes the name lock", field, clearable.Column)
+			}
+		}
+		// Without this the loop above proves nothing on the day no name is
+		// clearable any more — every case would fall through the non-name arm
+		// and the test would pass over a question it never asked.
+		if names == 0 {
+			t.Fatal("no clearable field writes a name column, so this test checked nothing")
+		}
+	})
+
+	t.Run("an edit that touches no name", func(t *testing.T) {
+		owner := "somebody"
+		if renamesAnOrganization(UpdateOrganizationInput{
+			Industry: &owner, Clear: []string{"description", "industry"},
+		}) {
+			t.Error("an edit writing no name column took the workspace-wide lock")
+		}
+	})
 }
